fix: Namespaced PlannedWorks
