--- app/graphql/types/planned_work_type.rb
@@ -1,8 +1,0 @@
-module Types
-  class PlannedWorkType < Types::BaseObject
-    field :createdDateTime, String, null: true
-    field :description, String, null: true
-    field :id, String, null: true
-    field :lastUpdateDateTime, String, null: true
-  end
-end
