+++ app/graphql/types/tfl/entities/journey_planner/planned_work_type.rb
@@ -1,0 +1,14 @@
+module Types
+  module Tfl
+    module Entities
+      module JourneyPlanner
+        class PlannedWorkType < Types::BaseObject
+          field :createdDateTime, String, null: true
+          field :description, String, null: true
+          field :id, String, null: true
+          field :lastUpdateDateTime, String, null: true
+        end
+      end
+    end
+  end
+end
